@@ -1,7 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import reportWebVitals from './reportWebVitals';
 import { mergeStyles } from '@fluentui/react';
 
 // Application Insights - https://github.com/microsoft/ApplicationInsights-JS/tree/master/extensions/applicationinsights-react-js
@@ -13,13 +14,15 @@ import baseConfig from "./config.json";
 import principalConfig from "./configpresets/principals.json"
 import entScaleConfig from "./configpresets/entScale.json"
 import baselineConfig from "./configpresets/baselines.json"
+import gamingConfig from "./configpresets/gaming.json"
 
 const configData = {
   ...baseConfig,
   presets: {
     ...principalConfig,
     ...entScaleConfig,
-    ...baselineConfig
+    ...baselineConfig,
+    ...gamingConfig
   }
 };
 
@@ -41,11 +44,16 @@ mergeStyles({
   },
 });
 
-
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-      <ConfigContext.Provider value={configData}>
+    <ConfigContext.Provider value={configData}>
         <App />
       </ConfigContext.Provider>
-      </React.StrictMode>, document.getElementById('root'))
+  </React.StrictMode>
+);
 
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
